处理类型转换异常

+++ lib/flutter_umeng_analytics.dart
@@ -17,7 +17,7 @@ class UMengAnalytics {
       {int policy,
       bool reportCrash,
       bool encrypt,
-      double interval,
+      int interval,
       String channel,
       bool logEnable}) {
     Map<String, dynamic> args = {"key": key};
